+++ ESI.NET/Models/Calendar/Response.cs
@@ -6,7 +6,7 @@ namespace ESI.NET.Models.Calendar
     public class Response
     {
         [JsonProperty("character_id")]
-        public int CharacterId { get; set; }
+        public long CharacterId { get; set; }
 
         [JsonProperty("event_response")]
         public EventResponse EventResponse { get; set; }
