+++ ESI.NET/Models/Character/Agent.cs
@@ -6,7 +6,7 @@ namespace ESI.NET.Models.Character
     public class Agent
     {
         [JsonProperty("agent_id")]
-        public int AgentId { get; set; }
+        public long AgentId { get; set; }
 
         [JsonProperty("points_per_day")]
         public float PointsPerDay { get; set; }
@@ -15,7 +15,7 @@ namespace ESI.NET.Models.Character
         public float RemainderPoints { get; set; }
 
         [JsonProperty("skill_type_id")]
-        public int SkillTypeId { get; set; }
+        public long SkillTypeId { get; set; }
 
         [JsonProperty("started_at")]
         public DateTime StartedAt { get; set; }
